Move TokenColumnsFeature
